Adding governance tx processing

+++ app/queries/responses/blockchain/ProposalDepositResponse.scala
@@ -8,7 +8,7 @@ import transactions.Abstract.BaseResponse
 object ProposalDepositResponse {
 
   case class Deposit(proposal_id: String, depositor: String, amount: Seq[Coin]) {
-    def toSerializableProposalDeposit: BlockchainProposalDeposit = BlockchainProposalDeposit(proposalID = proposal_id, depositor = depositor, amount = amount.map(_.toCoin))
+    def toSerializableProposalDeposit: BlockchainProposalDeposit = BlockchainProposalDeposit(proposalID = proposal_id.toInt, depositor = depositor, amount = amount.map(_.toCoin))
   }
 
   implicit val depositReads: Reads[Deposit] = Json.reads[Deposit]
